Data Analysis done on cleaned property data
The data had been cleaned and had visual statistics created for it. Now a number of analytics were done to understand the relationships between customers age and price of property. The following analytics were found: The number of houses sold based on age intervals, number of houses sold/not sold based on price interval, relationship between age and price (covariance and correlation.